Edited ServingController.cs
- Player is now able to serve dishes to customers seated at the table
- Cleaned up some of the code accordingly
