+++ CULLinary/Assets/Experiment/TestAssets/Restaurant/Restaurant Animations/Restaurant_Player_Animator.controller.meta
