+++ CULLinary/Assets/Experiment/TestScripts/Restaurant Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 491087e5ec14b5c4480b6cf16550c473
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
