+++ CULLinary/Assets/Experiment/TestScripts/Restaurant Scripts/Frying.anim
