+++ CULLinary/ProjectSettings/TagManager.asset
@@ -7,6 +7,7 @@ TagManager:
   - PlayerBody
   - CookingStation
   - FoodToServe
+  - Customer
   layers:
   - Default
   - TransparentFX
